Restructured the modules. Prettified the code
