+++ src/modules/user/user.service.ts
@@ -16,7 +16,7 @@ import {
   UserLoginResponseDto,
   ChangePasswordDto,
   GetUsersQueryParams,
-} from '../dtos';
+} from '../../dtos';
 
 @Injectable()
 export class UserService {
